exp/comm/py_practice/django_demo/requirements.txt
django
djangorestframework
django-rest-knox
django-health-check